backend-drf/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.8.2
python-decouple==3.8
dj-database-url==2.2.0
